[image_picker] Check camera permissions and return error (#1440)
Under iOS, the camera permissions have not been correctly checked. If the user did not allow camera access, the camera view was launched and displayed a black screen. It was then also possible to make a photo of the black screen and use it. The camera view controller is now only launched, when the user has granted permissions. Otherwise it will not launch and flutter will also receive an error.
flutter/flutter#30378