[various] Update Android example min SDKs (#3847)
Updates all the example apps that had a `minSdkVersion` of 16 to instead use `flutter.compileSdkVersion`. That is 16 on stable, but now 19 on master since Flutter support for 16-18 is being fully removed.

This unblocks the roller, since example apps that depend on `integration_test` now require a minimum of 19+ on master.